Move networkx convert.py to `core.dag` module
