+++ test/unit/visualization/test_visualisation_utils.py
@@ -1,8 +1,8 @@
 from fedot.core.dag.graph_utils import distance_to_primary_level
+from fedot.core.dag.convert import graph_structure_as_nx_graph
 from fedot.core.optimisers.fitness.multi_objective_fitness import MultiObjFitness
 from fedot.core.optimisers.opt_history_objects.individual import Individual
 from fedot.core.pipelines.adapters import PipelineAdapter
-from fedot.core.pipelines.convert import graph_structure_as_nx_graph
 from fedot.core.pipelines.node import PrimaryNode, SecondaryNode
 from fedot.core.pipelines.pipeline import Pipeline
 from fedot.core.visualisation.graph_viz import get_hierarchy_pos
